TemplateScripts finder and executor
Instead of the rke-openstack tool knowing about all the stages and
stuff. We plan on having the rega tool find the scripts in the template
repo automatically. That requires there to be a scripts/ directory in
the template repo within which there are a number of scripts following
the following format:

  scripts/<stage>_<type>_<name>.sh

Where stage is a number telling the tool and others what order to run
things in and <type> is a type. Planned types that will have special
meaning to rega are init, apply, destroy and plan. They all conform to
their respective command in the tool. <name> is just a stage name that
should have meaning to the person running the scripts.