Pull the master branch of the templates
If we always pull the branch/tag with the same version of the tool we
are tightly coupling the different parts to much and it will make it
tricky to do updates to one or the other.

Maybe we should consider using semantic versioning here somehow and try
to figure out how to get the latest branch that has the same major
version as the rega tool repo.